+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 2021-31-01
+### Added
+- added `nested` support for `folio_console_controller_for`
+
 ## 2021-20-01
 ### Changed
 - changed api files controllers to use direct s3 upload